Delete all P2PClient stuff
