--- node/src/main/scala/org/bitcoins/node/networking/P2PClient.scala
@@ -1,754 +1,0 @@
-package org.bitcoins.node.networking
-
-import akka.actor.{Actor, ActorRef, ActorSystem, Props, Terminated}
-import akka.event.LoggingReceive
-import akka.io.Tcp.SO.KeepAlive
-import akka.io.{IO, Tcp}
-import akka.pattern.ask
-import akka.util.{ByteString, CompactByteString, Timeout}
-import org.bitcoins.chain.config.ChainAppConfig
-import org.bitcoins.core.config.NetworkParameters
-import org.bitcoins.core.p2p.{
-  ExpectsResponse,
-  NetworkHeader,
-  NetworkMessage,
-  NetworkPayload
-}
-import org.bitcoins.core.util.NetworkUtil
-import org.bitcoins.node.config.NodeAppConfig
-import org.bitcoins.node.models.Peer
-import org.bitcoins.node.networking.P2PClient.{
-  ExpectResponseCommand,
-  NodeCommand
-}
-import org.bitcoins.node.networking.peer.PeerMessageReceiver.NetworkMessageReceived
-import org.bitcoins.node.networking.peer.PeerMessageReceiverState._
-import org.bitcoins.node.networking.peer.{
-  PeerMessageReceiver,
-  PeerMessageReceiverState
-}
-import org.bitcoins.node.util.PeerMessageSenderApi
-import org.bitcoins.node.{P2PLogger, ResponseTimeout}
-import org.bitcoins.tor.Socks5Connection.{Socks5Connect, Socks5Connected}
-import org.bitcoins.tor.{Socks5Connection, Socks5ProxyParams}
-import scodec.bits.ByteVector
-
-import java.net.InetSocketAddress
-import scala.annotation.tailrec
-import scala.concurrent.duration.DurationInt
-import scala.concurrent.{Await, ExecutionContext, Future}
-import scala.util._
-
-/** This actor is responsible for creating a connection,
-  * relaying messages and closing a connection to our peer on
-  * the P2P network. This is the actor that directly interacts
-  * with the p2p network. It's responsibly is to deal with low
-  * level .TCP messages.
-  *
-  * If the client receives a [[org.bitcoins.core.p2p.NetworkMessage NetworkMessage]], from a
-  * [[org.bitcoins.node.networking.peer.PeerMessageSender PeerMessageSender]]
-  * it serializes the message to it to a [[akka.util.ByteString]] and then
-  * sends it to the internal `manager` which streams the data to our peer
-  * on the Bitcoin network.
-  *
-  * If the client receives a [[Tcp.Received]] message, it means we have received
-  * a message from our peer on the Bitcoin P2P network. This means we try to parse
-  * the bytes into a [[org.bitcoins.core.p2p.NetworkMessage NetworkMessage]].
-  * If we successfully parse the message we relay that message to the
-  * [[org.bitcoins.node.networking.peer.PeerMessageSender PeerMessageSender]]
-  * that created the Client Actor.
-  *
-  * In this class you will see a 'unalignedBytes' value passed around in a lot of methods.
-  * This is because we cannot assume that a Bitcoin P2P message aligns with a TCP packet.
-  * For instance, a large block message (up to 4MB in size)
-  * CANNOT fit in a single TCP packet. This means we must cache
-  * the bytes and wait for the rest of them to be sent.
-  *
-  * @param initPeerMsgHandlerReceiver The place we send messages that we successfully parsed
-  *                               from our peer on the P2P network. This is mostly likely
-  *                               a [[org.bitcoins.node.networking.peer.PeerMessageSender]]
-  */
-case class P2PClientActor(
-    peer: Peer,
-    initPeerMessageReceiver: PeerMessageReceiver,
-    peerMessageSenderApi: PeerMessageSenderApi,
-    maxReconnectionTries: Int
-)(implicit nodeAppConfig: NodeAppConfig, chainAppConfig: ChainAppConfig)
-    extends Actor
-    with P2PLogger {
-
-  import context.dispatcher
-
-  private var currentPeerMessageReceiver = initPeerMessageReceiver
-
-  private var reconnectHandlerOpt: Option[Peer => Future[Unit]] = None
-
-  //total
-  private var reconnectionTry = 0
-
-  //from the last initialization
-  private var curReconnectionTry = 0
-
-  private val reconnectionDelay = 500.millis
-
-  /** The parameters for the network we are connected to
-    */
-  private val network: NetworkParameters = nodeAppConfig.network
-
-  private val timeout = 10.seconds
-
-  /** The manager is an actor that handles the underlying low level I/O resources (selectors, channels)
-    * and instantiates workers for specific tasks, such as listening to incoming connections.
-    */
-  def manager: ActorRef = IO(Tcp)(context.system)
-
-  /** TODO: this comment seems wrong?
-    *
-    * This actor signifies the node we are connected to on the p2p network
-    * This is the context we are in after we received a [[Tcp.Connected]] message
-    */
-  private def awaitNetworkRequest(
-      peerConnection: ActorRef,
-      unalignedBytes: ByteVector): Receive =
-    LoggingReceive {
-      case message: NetworkMessage =>
-        message.payload match {
-          case _: ExpectsResponse =>
-            logger.debug(
-              s"${message.payload.commandName} expects response from $peer")
-            Await.result(handleExpectResponse(message.payload), timeout)
-          case _ =>
-        }
-        sendNetworkMessage(message, peerConnection)
-      case ResponseTimeout(msg) =>
-        currentPeerMessageReceiver = Await.result(
-          currentPeerMessageReceiver.onResponseTimeout(networkPayload = msg,
-                                                       peer = peer),
-          timeout)
-      case payload: NetworkPayload =>
-        val networkMsg = NetworkMessage(network, payload)
-        self.forward(networkMsg)
-      case message: Tcp.Event =>
-        val newUnalignedBytes =
-          handleEvent(message, peerConnection, unalignedBytes)
-        context.become(awaitNetworkRequest(peerConnection, newUnalignedBytes))
-      case P2PClient.CloseCommand =>
-        handleNodeCommand(cmd = P2PClient.CloseCommand,
-                          peerConnectionOpt = Some(peerConnection))
-      case P2PClient.CloseAnyStateCommand =>
-        handleNodeCommand(cmd = P2PClient.CloseAnyStateCommand,
-                          peerConnectionOpt = Some(peerConnection))
-      case metaMsg: P2PClient.MetaMsg =>
-        sender() ! handleMetaMsg(metaMsg)
-
-      case ExpectResponseCommand(msg) =>
-        Await.result(handleExpectResponse(msg), timeout)
-      case Terminated(actor) if actor == peerConnection =>
-        reconnect()
-
-      case networkMessageReceived: NetworkMessageReceived =>
-        val newMsgReceiverStateF =
-          handleReceivedMsgFn(networkMessageReceived)
-        val newMsgReceiverState =
-          try {
-            Await.result(newMsgReceiverStateF, timeout)
-          } catch {
-            case scala.util.control.NonFatal(err) =>
-              logger.error(
-                s"Failed to process message in time state=${currentPeerMessageReceiver.state}, msgs=${networkMessageReceived.msg.header.commandName}",
-                err)
-              throw err
-          }
-        currentPeerMessageReceiver = newMsgReceiverState
-        if (currentPeerMessageReceiver.state.isInitialized) {
-          curReconnectionTry = 0
-          reconnectHandlerOpt.foreach(_.apply(peer))
-          reconnectHandlerOpt = None
-        }
-    }
-
-  /** Behavior to ignore network messages. Used only when the peer is being disconnected by us as we would not want to
-    * process any messages from it in that state.
-    */
-  private def ignoreNetworkMessages(
-      peerConnectionOpt: Option[ActorRef],
-      unalignedBytes: ByteVector): Receive = LoggingReceive {
-    case msg @ (_: NetworkMessage | _: NetworkPayload |
-        _: ExpectResponseCommand) =>
-      logger.debug(s"Ignoring $msg for $peer as disconnecting.")
-    case message: Tcp.Event if peerConnectionOpt.isDefined =>
-      val newUnalignedBytes =
-        handleEvent(message, peerConnectionOpt.get, unalignedBytes)
-      context.become(
-        ignoreNetworkMessages(peerConnectionOpt, newUnalignedBytes))
-    case _ @(P2PClient.CloseCommand | P2PClient.CloseAnyStateCommand) =>
-    //ignore
-    case metaMsg: P2PClient.MetaMsg =>
-      sender() ! handleMetaMsg(metaMsg)
-    case Terminated(actor)
-        if peerConnectionOpt.isDefined && actor == peerConnectionOpt.get =>
-      reconnect()
-  }
-
-  override def receive: Receive = LoggingReceive {
-    case () =>
-      sender() ! ()
-    case cmd: NodeCommand =>
-      handleNodeCommand(cmd = cmd, peerConnectionOpt = None)
-    case metaMsg: P2PClient.MetaMsg =>
-      sender() ! handleMetaMsg(metaMsg)
-  }
-
-  def reconnecting: Receive = LoggingReceive {
-    case P2PClient.ReconnectCommand =>
-      logger.debug(s"reconnecting to ${peer.socket}")
-      connect()
-    case P2PClient.CloseCommand =>
-      handleNodeCommand(P2PClient.CloseCommand, None)
-    case P2PClient.ConnectCommand =>
-      handleNodeCommand(P2PClient.ConnectCommand, None)
-    case P2PClient.CloseAnyStateCommand =>
-      handleNodeCommand(cmd = P2PClient.CloseAnyStateCommand,
-                        peerConnectionOpt = None)
-    case msg: NetworkMessage =>
-      logger.debug(s"$peer got ${msg.payload.commandName} while reconnecting.")
-      if (msg.payload.isInstanceOf[ExpectsResponse])
-        Await.result(handleExpectResponse(msg.payload), timeout)
-    case ExpectResponseCommand(msg) =>
-      Await.result(handleExpectResponse(msg), timeout)
-    case metaMsg: P2PClient.MetaMsg =>
-      sender() ! handleMetaMsg(metaMsg)
-  }
-
-  def connecting(proxyParams: Option[Socks5ProxyParams]): Receive = {
-    LoggingReceive {
-      case P2PClient.CloseAnyStateCommand =>
-        handleNodeCommand(cmd = P2PClient.CloseAnyStateCommand,
-                          peerConnectionOpt = None)
-      case msg: NetworkMessage =>
-        if (msg.payload.isInstanceOf[ExpectsResponse])
-          Await.result(handleExpectResponse(msg.payload), timeout)
-      case ExpectResponseCommand(msg) =>
-        Await.result(handleExpectResponse(msg), timeout)
-      case Tcp.CommandFailed(c: Tcp.Connect) =>
-        val peerOrProxyAddress = c.remoteAddress
-        logger.debug(
-          s"connection failed to ${peerOrProxyAddress} ${proxyParams}")
-        reconnect()
-
-      case event @ Tcp.Connected(peerOrProxyAddress, _) =>
-        val connection = sender()
-        proxyParams match {
-          case Some(proxyParams) =>
-            val proxyAddress = peerOrProxyAddress
-            val remoteAddress = peer.socket
-            logger.debug(s"connected to SOCKS5 proxy ${proxyAddress}")
-            logger.debug(
-              s"connecting to ${remoteAddress} via SOCKS5 ${proxyAddress}")
-            val proxy =
-              context.actorOf(Socks5Connection.props(
-                                sender(),
-                                Socks5ProxyParams.proxyCredentials(proxyParams),
-                                Socks5Connect(remoteAddress)),
-                              "Socks5Connection")
-            context watch proxy
-            context become socks5Connecting(event,
-                                            proxy,
-                                            remoteAddress,
-                                            proxyAddress)
-          case None =>
-            val peerAddress = peerOrProxyAddress
-            logger.debug(s"connected to ${peerAddress}")
-            context watch connection
-            val _ = handleEvent(event, connection, ByteVector.empty)
-        }
-      case metaMsg: P2PClient.MetaMsg =>
-        sender() ! handleMetaMsg(metaMsg)
-    }
-  }
-
-  def socks5Connecting(
-      event: Tcp.Connected,
-      proxy: ActorRef,
-      remoteAddress: InetSocketAddress,
-      proxyAddress: InetSocketAddress): Receive = LoggingReceive {
-    case P2PClient.CloseAnyStateCommand =>
-      handleNodeCommand(cmd = P2PClient.CloseAnyStateCommand,
-                        peerConnectionOpt = None)
-    case msg: NetworkMessage =>
-      if (msg.payload.isInstanceOf[ExpectsResponse])
-        Await.result(handleExpectResponse(msg.payload), timeout)
-    case ExpectResponseCommand(msg) =>
-      Await.result(handleExpectResponse(msg), timeout)
-    case Tcp.CommandFailed(_: Socks5Connect) =>
-      logger.debug(
-        s"connection failed to ${remoteAddress} via SOCKS5 ${proxyAddress}")
-      reconnect()
-    case Socks5Connected(_) =>
-      logger.debug(
-        s"connected to ${remoteAddress} via SOCKS5 proxy ${proxyAddress}")
-      val _ = handleEvent(event, proxy, ByteVector.empty)
-    case Terminated(actor) if actor == proxy =>
-      reconnect()
-    case metaMsg: P2PClient.MetaMsg =>
-      sender() ! handleMetaMsg(metaMsg)
-  }
-
-  override def unhandled(message: Any): Unit = message match {
-    case payload: NetworkPayload =>
-      logger.debug(
-        s"Cannot send a message to our peer when we are not connected! payload=${payload} peer=${peer}")
-    case _ =>
-      logger.debug(s"unhandled message=$message")
-  }
-
-  private def connect(): Unit = {
-    val (peerOrProxyAddress, proxyParams) =
-      peer.socks5ProxyParams match {
-        case Some(proxyParams) =>
-          val host = peer.socket.getHostString
-          if (!NetworkUtil.isLocalhost(host)) {
-            val proxyAddress = proxyParams.address
-            logger.debug(s"connecting to SOCKS5 proxy $proxyAddress")
-            (proxyAddress, Some(proxyParams))
-          } else {
-            val remoteAddress = peer.socket
-            logger.debug(s"connecting to $remoteAddress")
-            (peer.socket, None)
-          }
-        case None =>
-          val remoteAddress = peer.socket
-          logger.debug(s"connecting to $remoteAddress")
-          (peer.socket, None)
-      }
-    val options = Vector(KeepAlive(true))
-    manager ! Tcp.Connect(peerOrProxyAddress,
-                          timeout = Some(nodeAppConfig.connectionTimeout),
-                          options = options,
-                          pullMode = true)
-
-    context become connecting(proxyParams)
-  }
-
-  private def reconnect(): Unit = {
-    currentPeerMessageReceiver.state match {
-      case _: PeerMessageReceiverState.InitializedDisconnect |
-          _: PeerMessageReceiverState.InitializedDisconnectDone =>
-        logger.debug(
-          s"Ignoring reconnection attempts as we initialized disconnect from peer=$peer, state=${currentPeerMessageReceiver.state}")
-        context.stop(self)
-      case bad: PeerMessageReceiverState.StoppedReconnect =>
-        throw new RuntimeException(s"Tried to reconnect when in state $bad")
-      case state @ (PeerMessageReceiverState.Preconnection | _: Initializing |
-          _: Normal | _: Disconnected | _: Waiting) =>
-        state match {
-          case wait: Waiting =>
-            currentPeerMessageReceiver = Await.result(
-              currentPeerMessageReceiver.onResponseTimeout(networkPayload =
-                                                             wait.responseFor,
-                                                           peer = peer),
-              timeout)
-          case init: Initializing =>
-            init.initializationTimeoutCancellable.cancel()
-          case _ =>
-        }
-
-        logger.debug(
-          s"Attempting to reconnect to peer=$peer, tryCount=$reconnectionTry, previous state=${currentPeerMessageReceiver.state}")
-
-        if (reconnectionTry >= maxReconnectionTries) {
-          logger.debug("Exceeded maximum number of reconnection attempts")
-          context.stop(self)
-        } else {
-          val delay = reconnectionDelay * (1 << curReconnectionTry)
-          curReconnectionTry += 1
-          reconnectionTry = reconnectionTry + 1
-
-          import context.dispatcher
-          context.system.scheduler.scheduleOnce(delay)(
-            self ! P2PClient.ReconnectCommand)
-
-          context.become(reconnecting)
-        }
-    }
-  }
-
-  /** This function is responsible for handling a [[Tcp.Event]] algebraic data type
-    */
-  private def handleEvent(
-      event: Tcp.Event,
-      peerConnection: ActorRef,
-      unalignedBytes: ByteVector): ByteVector = {
-    event match {
-      case Tcp.Bound(localAddress) =>
-        logger.debug(
-          s"Actor is now bound to the local address: ${localAddress}")
-        context.parent ! Tcp.Bound(localAddress)
-
-        unalignedBytes
-      case Tcp.CommandFailed(command) =>
-        logger.debug(s"Client Command failed: ${command}")
-
-        unalignedBytes
-      case Tcp.Connected(remote, local) =>
-        logger.debug(s"Tcp connection to: ${remote}")
-        logger.debug(s"Local: ${local}")
-
-        //this is what registers a actor to send all byte messages to that is
-        //received from our peer. Since we are using 'self' that means
-        //our bitcoin peer will send all messages to this actor.
-        peerConnection ! Tcp.Register(self)
-        peerConnection ! Tcp.ResumeReading
-        val client = P2PClient(self, peer)
-        currentPeerMessageReceiver = currentPeerMessageReceiver.connect(
-          client,
-          peerMessageSenderApi)(context.system, nodeAppConfig, chainAppConfig)
-        context.become(awaitNetworkRequest(peerConnection, unalignedBytes))
-        unalignedBytes
-
-      case Tcp.ErrorClosed(cause) =>
-        logger.debug(
-          s"An error occurred in our connection with $peer, cause=$cause state=${currentPeerMessageReceiver.state}")
-        currentPeerMessageReceiver = Await.result(
-          currentPeerMessageReceiver.disconnect(peer)(context.system),
-          timeout)
-        context.stop(self)
-        unalignedBytes
-      case closeCmd @ (Tcp.ConfirmedClosed | Tcp.Closed | Tcp.Aborted |
-          Tcp.PeerClosed) =>
-        logger.info(
-          s"We've been disconnected by $peer command=${closeCmd} state=${currentPeerMessageReceiver.state}")
-        currentPeerMessageReceiver = Await.result(
-          currentPeerMessageReceiver.disconnect(peer)(context.system),
-          timeout)
-
-        context.stop(self)
-        unalignedBytes
-
-      case Tcp.Received(byteString: ByteString) =>
-        val byteVec = ByteVector(byteString.toArray)
-        logger.debug(s"Received ${byteVec.length} TCP bytes")
-        logger.trace(s"Received TCP bytes: ${byteVec.toHex}")
-        logger.trace {
-          val post =
-            if (unalignedBytes.isEmpty) "None"
-            else unalignedBytes.toHex
-          s"Unaligned bytes: $post"
-        }
-
-        if (unalignedBytes.isEmpty) {
-          peerConnection ! Tcp.ResumeReading
-        }
-
-        //we need to aggregate our previous 'unalignedBytes' with the new message
-        //we just received from our peer to hopefully be able to parse full messages
-        val bytes: ByteVector = unalignedBytes ++ byteVec
-        logger.trace(s"Bytes for message parsing: ${bytes.toHex}")
-        val (messages, newUnalignedBytes) =
-          P2PClient.parseIndividualMessages(bytes)
-
-        logger.debug {
-          val length = messages.length
-          val suffix = if (length == 0) "" else s": ${messages.mkString(", ")}"
-
-          s"Parsed $length message(s) from bytes$suffix"
-        }
-        logger.debug(s"Unaligned bytes after this: ${newUnalignedBytes.length}")
-        if (newUnalignedBytes.nonEmpty) {
-          logger.trace(s"Unaligned bytes: ${newUnalignedBytes.toHex}")
-        }
-
-        logger.trace(s"About to process ${messages.length} messages")
-        messages.foreach(m => self ! NetworkMessageReceived(m, peer))
-        peerConnection ! Tcp.ResumeReading
-        newUnalignedBytes
-    }
-  }
-
-  private val handleReceivedMsgFn: NetworkMessageReceived => Future[
-    PeerMessageReceiver] = { case msg: NetworkMessageReceived =>
-    val resultF = if (currentPeerMessageReceiver.state.isConnected) {
-      currentPeerMessageReceiver.state match {
-        case _ @(_: Normal | _: Waiting | Preconnection | _: Initializing) =>
-          currentPeerMessageReceiver.handleNetworkMessageReceived(
-            msg,
-            peerMessageSenderApi)
-        case _: Disconnected | _: InitializedDisconnectDone |
-            _: InitializedDisconnect | _: StoppedReconnect =>
-          logger.debug(
-            s"Ignoring ${msg.msg.payload.commandName} from $peer as in state=${currentPeerMessageReceiver.state}")
-          Future.successful(currentPeerMessageReceiver)
-      }
-    } else {
-      Future.successful(currentPeerMessageReceiver)
-    }
-    resultF
-  }
-
-  /** Returns the current state of our peer given the [[P2PClient.MetaMsg meta message]]
-    */
-  private def handleMetaMsg(metaMsg: P2PClient.MetaMsg): Boolean = {
-    metaMsg match {
-      case P2PClient.IsConnected => currentPeerMessageReceiver.state.isConnected
-      case P2PClient.IsInitialized =>
-        currentPeerMessageReceiver.state.isInitialized
-      case P2PClient.IsDisconnected =>
-        currentPeerMessageReceiver.state.isDisconnected
-    }
-  }
-
-  /** Sends a network request to our peer on the network
-    */
-  private def sendNetworkMessage(
-      message: NetworkMessage,
-      peerConnection: ActorRef): Unit = {
-
-    val byteMessage = CompactByteString(message.bytes.toArray)
-    peerConnection ! Tcp.Write(byteMessage)
-    peerConnection ! Tcp.ResumeReading
-  }
-
-  /** Handles the commands that can be sent to our node
-    * @param cmd the command to execute
-    * @param peerConnectionOpt the connection to a peer, if it exists
-    */
-  private def handleNodeCommand(
-      cmd: NodeCommand,
-      peerConnectionOpt: Option[ActorRef]): Unit = {
-    cmd match {
-      case P2PClient.ConnectCommand =>
-        connect()
-      case P2PClient.ReconnectCommand =>
-        reconnect()
-      case P2PClient.CloseCommand =>
-        peerConnectionOpt match {
-          case Some(peerConnection) =>
-            logger.info(s"Disconnecting from peer $peer")
-            context become ignoreNetworkMessages(Some(peerConnection),
-                                                 ByteVector.empty)
-            currentPeerMessageReceiver =
-              currentPeerMessageReceiver.initializeDisconnect(peer)
-            peerConnection ! Tcp.Close
-          case None =>
-            logger.warn(
-              s"Attempting to disconnect peer that was not connected!")
-        }
-      case P2PClient.CloseAnyStateCommand =>
-        logger.info(
-          s"Received close any state for peer=$peer peerConnectionOpt=$peerConnectionOpt currentPeerMsgRecvState=${currentPeerMessageReceiver.state}")
-        peerConnectionOpt match {
-          case Some(peerConnection) =>
-            context become ignoreNetworkMessages(Some(peerConnection),
-                                                 ByteVector.empty)
-            currentPeerMessageReceiver =
-              currentPeerMessageReceiver.initializeDisconnect(peer)
-            peerConnection ! Tcp.Close
-          case None =>
-            context become ignoreNetworkMessages(None, ByteVector.empty)
-            currentPeerMessageReceiver = Await.result(
-              currentPeerMessageReceiver.stopReconnect(peer),
-              timeout)
-            context.stop(self)
-        }
-    }
-  }
-
-  /** For any [[org.bitcoins.core.p2p.NetworkPayload]], if it a subtype of [[ExpectsResponse]], starts a
-    * cancellable timer and changes state to [[PeerMessageReceiverState.Waiting]]. Note that any messages, received
-    * while waiting for a particular message are still processed, they just won't cancel the scheduled query timeout.
-    * Can only wait for one query at a time, other messages that are [[ExpectsResponse]] and received while in
-    * [[PeerMessageReceiverState.Waiting]] are still sent, but no query timeout would be executed for those.
-    * Currently, such a situation is not meant to happen.
-    */
-  private def handleExpectResponse(msg: NetworkPayload): Future[Unit] = {
-    require(
-      msg.isInstanceOf[ExpectsResponse],
-      s"Tried to wait for response to message which is not a query, got=$msg")
-    logger.debug(s"Expecting response for ${msg.commandName} for $peer")
-    currentPeerMessageReceiver
-      .handleExpectResponse(msg = msg, peer = peer)(context.system,
-                                                    nodeAppConfig)
-      .map { newReceiverState =>
-        currentPeerMessageReceiver = newReceiverState
-      }
-  }
-}
-
-case class P2PClient(actor: ActorRef, peer: Peer) extends P2PLogger {
-  import akka.pattern.ask
-
-  def isConnected()(implicit
-      timeout: Timeout,
-      ec: ExecutionContext): Future[Boolean] = {
-    val isConnectedF = actor.ask(P2PClient.IsConnected).mapTo[Boolean]
-    isConnectedF.recoverWith { case _: Throwable =>
-      Future.successful(false)
-    }
-  }
-
-  def isInitialized()(implicit
-      timeout: Timeout,
-      ec: ExecutionContext): Future[Boolean] = {
-    val isInitF = actor.ask(P2PClient.IsInitialized).mapTo[Boolean]
-    isInitF.recoverWith { case _: Throwable =>
-      Future.successful(false)
-    }
-  }
-
-  def isDisconnected()(implicit
-      timeout: Timeout,
-      ec: ExecutionContext): Future[Boolean] = {
-    val isDisconnect: Future[Boolean] =
-      actor.ask(P2PClient.IsDisconnected).mapTo[Boolean]
-
-    //this future can be failed, as we stop the P2PClientActor if we send a disconnect
-    //if that actor has been killed, the peer _has_ to have been disconnected
-    isDisconnect.recoverWith { case _: Throwable =>
-      Future.successful(true)
-    }
-  }
-
-  /** Closes the P2P client
-    */
-  def close(): Unit = {
-    actor ! P2PClient.CloseAnyStateCommand
-  }
-}
-
-object P2PClient extends P2PLogger {
-
-  sealed trait NodeCommand
-  case object ConnectCommand extends NodeCommand
-  case object ReconnectCommand extends NodeCommand
-  case object CloseCommand extends NodeCommand
-  case object CloseAnyStateCommand extends NodeCommand
-
-  case class ExpectResponseCommand(msg: NetworkPayload)
-
-  /** A message hierarchy that canbe sent to [[P2PClientActor P2P Client Actor]]
-    * to query about meta information of a peer
-    */
-  sealed trait MetaMsg
-
-  /** A message that can be sent to [[P2PClient p2p client]] that returns true
-    * if the peer is connected, false if not
-    */
-  final case object IsConnected extends MetaMsg
-
-  /** A message that can be sent to [[P2PClient p2p client]] that returns true
-    * if the peer is initialized (p2p handshake complete), false if not
-    */
-  final case object IsInitialized extends MetaMsg
-
-  /** A message that can be sent to [[P2PClient p2p client]] that returns true
-    * if the peer is disconnected, false otherwise
-    */
-  final case object IsDisconnected extends MetaMsg
-
-  def props(
-      peer: Peer,
-      peerMsgHandlerReceiver: PeerMessageReceiver,
-      peerMessageSenderApi: PeerMessageSenderApi,
-      maxReconnectionTries: Int)(implicit
-      nodeAppConfig: NodeAppConfig,
-      chainAppConfig: ChainAppConfig
-  ): Props = {
-    Props(
-      classOf[P2PClientActor],
-      peer,
-      peerMsgHandlerReceiver,
-      peerMessageSenderApi,
-      maxReconnectionTries,
-      nodeAppConfig,
-      chainAppConfig
-    )
-  }
-
-  def apply(
-      peer: Peer,
-      peerMessageReceiver: PeerMessageReceiver,
-      peerMessageSenderApi: PeerMessageSenderApi,
-      maxReconnectionTries: Int = 16,
-      supervisor: ActorRef)(implicit
-      nodeAppConfig: NodeAppConfig,
-      chainAppConfig: ChainAppConfig,
-      system: ActorSystem): Future[P2PClient] = {
-    val clientProps = props(
-      peer = peer,
-      peerMsgHandlerReceiver = peerMessageReceiver,
-      peerMessageSenderApi,
-      maxReconnectionTries = maxReconnectionTries
-    )
-
-    import system.dispatcher
-    implicit val timeout: Timeout = Timeout(10.second)
-    val actorRefF = supervisor ? clientProps
-
-    actorRefF.map { actorRef =>
-      P2PClient(actorRef.asInstanceOf[ActorRef], peer)
-    }
-  }
-
-  /** Akka sends messages as one byte stream. There is not a 1 to 1 relationship between byte streams received and
-    * bitcoin protocol messages. This function parses our byte stream into individual network messages
-    *
-    * @param bytes the bytes that need to be parsed into individual messages
-    * @return the parsed [[NetworkMessage]]'s and the unaligned bytes that did not parse to a message
-    */
-  private[bitcoins] def parseIndividualMessages(
-      bytes: ByteVector): (Vector[NetworkMessage], ByteVector) = {
-    @tailrec
-    def loop(
-        remainingBytes: ByteVector,
-        accum: Vector[NetworkMessage]): (Vector[NetworkMessage], ByteVector) = {
-      if (remainingBytes.length <= 0) {
-        (accum, remainingBytes)
-      } else {
-        val headerTry = Try(
-          NetworkHeader.fromBytes(remainingBytes.take(NetworkHeader.bytesSize)))
-        headerTry match {
-          case Success(header) =>
-            val payloadBytes = remainingBytes
-              .drop(NetworkHeader.bytesSize)
-              .take(header.payloadSize.toInt)
-
-            val newRemainingBytes =
-              remainingBytes.drop(NetworkHeader.bytesSize + payloadBytes.size)
-
-            // If it's a message type we know, try to parse it
-            if (NetworkPayload.commandNames.contains(header.commandName)) {
-              Try(NetworkMessage(header.bytes ++ payloadBytes)) match {
-                case Success(message) =>
-                  logger.trace(
-                    s"Parsed a message=${message.header.commandName} from bytes, continuing with remainingBytes=${newRemainingBytes.length}")
-
-                  loop(newRemainingBytes, accum :+ message)
-                case Failure(_) =>
-                  // Can't parse message yet, we need to wait for more bytes
-                  (accum, remainingBytes)
-              }
-            } else if (payloadBytes.size == header.payloadSize.toInt) { // If we've received the entire unknown message
-              logger.debug(
-                s"Received unknown network message ${header.commandName}")
-              loop(newRemainingBytes, accum)
-            } else {
-              // If we can't parse the entire unknown message, continue on until we can
-              // so we properly skip it
-              (accum, remainingBytes)
-            }
-          case Failure(exc) =>
-            logger.trace(
-              s"Failed to parse network message $remainingBytes, could be because TCP frame isn't aligned: $exc")
-
-            //this case means that our TCP frame was not aligned with bitcoin protocol
-            //return the unaligned bytes so we can apply them to the next tcp frame of bytes we receive
-            //http://stackoverflow.com/a/37979529/967713
-            (accum, remainingBytes)
-        }
-      }
-    }
-    val (messages, remainingBytes) = loop(bytes, Vector.empty)
-    (messages, remainingBytes)
-  }
-
-}
